requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-docx==1.1.0
pydantic==2.5.3
google-genai==0.2.2
